Move RadioListItem out of BaseColorsSection dir
